Ajustes no exercício area

+++ Exercicios/area/src/main/java/br/com/gabrielferreira/model/AreaCalculavel.java
@@ -1,6 +1,8 @@
 package br.com.gabrielferreira.model;
 
+import java.math.BigDecimal;
+
 public interface AreaCalculavel {
 
-    Double calcularArea();
+    BigDecimal calcularArea();
 }
